+++ src/main/java/com/tishinserg/spreadbot/parsing/UnistreamRateParsingService.java
@@ -25,6 +25,7 @@ public class UnistreamRateParsingService implements ParsingService {
     private final WebClient uniRateParsingServiceWebClient;
     private final UnistreamRateParsingServiceProperties parsingServiceProperties;
 
+    //todo вынести хедеры в проперти
     @Override
     public Mono<Rate> getRate(String... params) {
         String countryName = params[0];